# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OneApp API
  version: 1.0.0
  description: |
    API contract for UK-first user registration, authentication, dashboard, and financial-planning features.
    Manual recurring-payment CRUD is deprecated while target-state recurring obligations move to bank-sourced detection.

    **Response envelope:** All API responses are wrapped in a standard envelope.
    Success responses return `{ "success": true, "data": { ... } }` where the `data` property contains the schema documented for each endpoint.
    Error responses return `{ "success": false, "error": "..." }`.
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /bank-sync/status:
    get:
      summary: Get mocked bank-link status
      description: Returns provider-strategy metadata, consent state, linked-account status, and the latest mocked ingestion summary for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Mocked bank-link status returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankSyncStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/mock-link:
    post:
      summary: Create mocked bank link
      description: Creates a mocked linked-account record for development and test environments without calling a live provider.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockBankLinkRequest'
      responses:
        '201':
          description: Mocked bank link created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankSyncStatusResponse'
        '400':
          description: Invalid mocked bank-link payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/mock-ingest:
    post:
      summary: Ingest mocked bank transactions
      description: Ingests mocked booked transactions using the internal linked-account contract, surfaces duplicate and partial-ingestion outcomes, and avoids any live provider dependency.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockBankIngestionRequest'
      responses:
        '200':
          description: Mocked bank transactions ingested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockBankIngestionResponse'
        '207':
          description: Mocked bank transactions ingested with duplicate or partial-ingestion warnings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockBankIngestionResponse'
        '400':
          description: Invalid mocked bank-ingestion payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/csv-import:
    post:
      summary: Import bank statement CSV
      description: Imports CSV-based bank statement rows and ingests them through the same internal transaction pipeline used by bank-linked ingestion.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CsvBankImportRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CsvBankImportUploadRequest'
      responses:
        '200':
          description: CSV statement ingested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockBankIngestionResponse'
        '207':
          description: CSV statement ingested with duplicate or partial-ingestion warnings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockBankIngestionResponse'
        '400':
          description: Invalid CSV import payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Uploaded CSV file exceeds the configured size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Uploaded multipart file is not a supported CSV-compatible file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/import-history:
    get:
      summary: List CSV import history
      description: Returns persisted CSV import summaries for the authenticated user ordered by most recent import first.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CSV import history returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsvImportHistoryResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/connect:
    post:
      summary: Initiate bank account linking
      description: Starts the OAuth consent flow with the configured UK Open Banking provider. Returns a consent URL that the client should redirect the user to.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Consent URL generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankSyncConnectResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Provider not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/callback:
    post:
      summary: Handle provider consent callback
      description: Exchanges the OAuth authorisation code from the provider callback for access and refresh tokens, discovers linked accounts, and persists consent state.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankSyncCallbackRequest'
      responses:
        '201':
          description: Bank account linked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankSyncCallbackResponse'
        '400':
          description: Invalid callback parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Provider rejected the authorisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/sync:
    post:
      summary: Trigger live transaction sync
      description: Fetches the latest transactions from the linked bank account via the provider API and ingests them through the standard transaction pipeline with deduplication.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Live sync completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveSyncResponse'
        '207':
          description: Live sync completed with partial success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveSyncResponse'
        '400':
          description: No active bank link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or expired bank link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Provider fetch failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Provider not configured or sync unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /bank-sync/webhook:
    post:
      summary: Receive provider webhook events
      description: Accepts provider transaction-update webhook events, validates the provider signature, deduplicates replayed events, and enqueues incremental sync jobs for matched linked accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankSyncWebhookEvent'
      responses:
        '202':
          description: Webhook accepted and a sync job was created or reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankSyncWebhookResponse'
        '200':
          description: Webhook acknowledged but ignored or deduplicated safely
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankSyncWebhookResponse'
        '400':
          description: Invalid webhook payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Webhook verification is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/disconnect:
    post:
      summary: Revoke bank account link
      description: Revokes the consent and clears stored provider tokens and linked account data. The user must re-consent to reconnect.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bank link revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankSyncDisconnectResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/jobs/enqueue:
    post:
      summary: Enqueue a sync job
      description: Creates a new sync job for the authenticated user. If a pending or running job already exists, returns the existing job without creating a duplicate.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncJobEnqueueRequest'
      responses:
        '201':
          description: New sync job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobEnqueueResponse'
        '200':
          description: Existing pending/running job returned (deduplicated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobEnqueueResponse'
        '400':
          description: No bank sync profile or invalid job type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/jobs/{jobId}/execute:
    post:
      summary: Execute a sync job
      description: Triggers execution of a pending or failed sync job. Fetches transactions from the provider, ingests them idempotently, and updates the sync checkpoint.
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sync job executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobExecuteResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job cannot be executed in its current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Provider error (transient or terminal)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/jobs/{jobId}:
    get:
      summary: Get sync job status
      description: Returns the current status and metadata of a sync job.
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobStatusResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bank-sync/jobs:
    get:
      summary: List sync job history
      description: Returns the authenticated user's sync job history, ordered by most recent first.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Job history returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobHistoryResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MockBankLinkRequest:
      type: object
      required:
        - accountId
        - accountName
        - sortCodeMasked
        - last4
      properties:
        accountId:
          type: string
        accountName:
          type: string
        sortCodeMasked:
          type: string
          description: Masked UK sort code in 00-00-00 format.
        last4:
          type: string
        linkedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...
    MockBankIngestionRequest:
      type: object
      required:
        - ingestionId
        - transactions
      properties:
        ingestionId:
          type: string
        receivedAt:
          type: string
          format: date-time
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/MockBankTransaction'
    CsvBankImportRequest:
      type: object
      required:
        - ingestionId
        - csvData
      description: |
        Supported CSV columns (aliases are accepted case-insensitively):
        - Required: `bookedAt` (`date`, `booked_at`, `transactionDate`), `amount` (`value`, `transactionAmount`), `merchantName` (`merchant`, `description`, `payee`)
        - Optional: `transactionId` (`transaction_id`, `id`, `reference`), `currency` (`ccy`), `direction` (`flow`, `type`), `status` (`transactionStatus`), `categoryHint` (`category`, `category_hint`)
        Parsing rules:
        - Negative amounts infer `direction=out`; positive amounts infer `direction=in` when direction is omitted.
        - `currency` defaults to `GBP`; `status` defaults to `booked`.
      properties:
        ingestionId:
          type: string
        receivedAt:
          type: string
          format: date-time
        csvData:
          type: string
          description: Raw CSV content including header row and transaction rows.
    CsvBankImportUploadRequest:
      type: object
      required:
        - ingestionId
        - file
      description: |
        Multipart upload variant of `CsvBankImportRequest`.
        - Accepted file MIME types: `text/csv`, `text/plain`, `application/vnd.ms-excel`
        - Maximum upload size: 1 MB
      properties:
        ingestionId:
          type: string
        receivedAt:
          type: string
          format: date-time
        file:
          type: string
          format: binary
...
    BankSyncStatusResponse:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/BankSyncMetadata'
        status:
          $ref: '#/components/schemas/BankSyncStatus'
    MockBankIngestionResponse:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/BankSyncMetadata'
        status:
          $ref: '#/components/schemas/BankSyncStatus'
        syncSummary:
          $ref: '#/components/schemas/SyncSummary'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/MockBankTransaction'
    CsvImportHistoryResponse:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/BankSyncMetadata'
        imports:
          type: array
          items:
            $ref: '#/components/schemas/SyncSummary'
    BankSyncConnectResponse:
      type: object
      properties:
        consentUrl:
          type: string
          description: URL to redirect the user to for provider consent.
        provider:
          type: string
        state:
          type: string
          description: Opaque state token for CSRF protection in the consent callback.
    BankSyncCallbackRequest:
      type: object
      required:
        - code
        - state
      properties:
        code:
          type: string
          description: OAuth authorisation code from the provider redirect.
        state:
          type: string
          description: State token originally issued by the connect endpoint.
    BankSyncCallbackResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/BankSyncStatusResponse'
        linked:
          type: boolean
          description: Whether an account was successfully discovered and linked.
        provider:
          type: string
        contractVersion:
          type: string
    LiveSyncResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/BankSyncStatusResponse'
        syncSummary:
          $ref: '#/components/schemas/SyncSummary'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/MockBankTransaction'
        transactionCount:
          type: integer
        contractVersion:
          type: string
...
    BankSyncWebhookEvent:
      type: object
      required:
        - eventId
        - type
        - accountId
      properties:
        eventId:
          type: string
        type:
          type: string
        accountId:
          type: string
        occurredAt:
          type: string
          format: date-time
...
    BankSyncDisconnectResponse:
      type: object
      properties:
        revoked:
          type: boolean
        provider:
          type: string
    SyncJobEnqueueRequest:
      type: object
      properties:
        jobType:
          type: string
          enum:
            - initial_sync
            - incremental_sync
            - retry_sync
          default: incremental_sync
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        details:
          type: object
          additionalProperties: true